fix(renderer): make the exact CPU path architecture-independent (Task 10.20)
The renderer's float64 arithmetic relied on the Go backend contracting each
multiply-add into one fused operation on ARM64. That matched the NEON span
kernel but matched neither amd64 nor the correctness oracle, which rounds with
math.Round and cannot fuse, so `internal/fit/renderer` was excluded from both
ARM64 rows of `ci-native-simd.yml`. Both rows now run it.

The recorded diagnosis was wrong on three counts, and the corrected one is in
`docs/known-limitations.md`:

- It named `compositeScalarPixel`, which does not exist; the alpha store is in
  `compositePixel`. Blocking that store alone changes nothing, because the
  oracle computes `outputAlpha` with the same expression shape and fuses with
  it.
- It said the failure needed ARM64 hardware. `qemu-aarch64-static` plus binfmt
  runs the cross-compiled test binary directly, reports NEON, and reproduces
  `pixel (4,11) channel 3 = 205, baseline = 206` exactly.
- It recorded one failing test where there are two. `TestPolishCircleBatchPool-
  WidthParity` also fails, against expected costs recorded on amd64.

Compiling the unmodified tree with `-gcflags=all=-d=fmahash=<pattern>` makes the
whole ARM64 suite pass, which bounds the cause to contraction alone, and
`bisect -compile=fma` named `composite_span.go:19:24` as the decisive site.

Two traps are worth naming. An explicit conversion rounds only what it wraps, so
`fgR + float64(bgR*bgBlend)` still fuses against `r*alpha` unless the
premultiplied foreground is rounded too. And the scalar span compositor was
deliberately fused to match the NEON kernel, so the kernel had to be unfused
with it: a 1:1 swap of `VMOV`+`VFMLA` for `FMUL`+`FADD`, hand-encoded because
Go's arm64 assembler has no vector mnemonic for them. Each encoding's comment is
what `go tool objdump` prints for it.

Nothing had ever compared that kernel against its scalar reference - both ARM64
span tests force the scalar tier, so they compared the scalar path with itself.
`TestCompositeSpanNEONMatchesScalar` now does, and its k/255 colour sampling is
load-bearing: uniformly random float colours found zero mismatches in 51.2
million evaluations against a reference that was genuinely wrong, where k/255
colours expose it in about half a percent of bytes. The test fails against the
previous fused kernel and passes against this one.

The span-geometry conversions are defensive rather than required: the suite
passes without them. They close the same class of divergence in the coverage
decision, once per row rather than per pixel.

amd64 output is unchanged - it emits no VFMADD before or after - and a 640x480
scene with NEON-length spans now renders to the same bytes and the same cost
under arm64/NEON and amd64/AVX2.

Verified: gofmt, go vet, go build, go test -short ./..., go test -race -short
./internal/fit/..., scripts/check-cross-build.sh (all six targets), golangci-lint
--new-from-merge-base=origin/main (0 issues), and the renderer and fit suites
under qemu-aarch64 at the NEON and forced-scalar tiers, including the
CIRCLEFIT_REQUIRE_SIMD_TIER combinations the workflow uses. ARM64 throughput was
not measured; emulated timings would not count, and `compositeSpanNEONMinPixels`
is unchanged.

+++ internal/fit/renderer/batch_polish.go
@@ -979,7 +979,7 @@ func circleRasterSpan(circle fit.Circle, y, width int) (int, int, bool) {
 
 	dy := float64(y) - circle.Y
 
-	remaining := circle.R*circle.R - dy*dy
+	remaining := float64(circle.R*circle.R) - float64(dy*dy)
 	if remaining < 0 {
 		return 0, 0, false
 	}
